Refactor preflight validations
