+++ pkg/validations/createvalidations/preflightvalidations.go
@@ -19,13 +19,13 @@ func (u *CreateValidations) PreflightValidations(ctx context.Context) (err error
 		{
 			Name:        "validate taints support",
 			Remediation: "ensure TAINTS_SUPPORT env variable is set",
-			Err:         ValidateTaintsSupport(u.Opts.Spec),
+			Err:         validations.ValidateTaintsSupport(u.Opts.Spec),
 			Silent:      true,
 		},
 		{
 			Name:        "validate node labels support",
 			Remediation: "ensure NODE_LABELS_SUPPORT env variable is set",
-			Err:         ValidateNodeLabelsSupport(u.Opts.Spec),
+			Err:         validations.ValidateNodeLabelsSupport(u.Opts.Spec),
 			Silent:      true,
 		},
 	}
@@ -56,17 +56,5 @@ func (u *CreateValidations) PreflightValidations(ctx context.Context) (err error
 		)
 	}
 
-	var errs []string
-	for _, validation := range createValidations {
-		if validation.Err != nil {
-			errs = append(errs, validation.Err.Error())
-		} else if !validation.Silent {
-			validation.LogPass()
-		}
-	}
-
-	if len(errs) > 0 {
-		return &validations.ValidationError{Errs: errs}
-	}
-	return nil
+	return validations.RunPreflightValidations(createValidations)
 }
